const vor class variables
