+++ src/datatypes/character/PropertyEnum.hpp
@@ -1,0 +1,38 @@
+//
+// Created by Carolin on 03.04.2020.
+//
+
+#ifndef LIBCOMMON_SRC_DATATYPES_CHARACTER_PROPERTYENUM_HPP_
+#define LIBCOMMON_SRC_DATATYPES_CHARACTER_PROPERTYENUM_HPP_
+
+
+namespace spy::character{
+  enum class PropertyEnum {
+    NIMBLENESS,
+    SLUGGISHNESS,
+    SPRYNESS,
+    AGILITY,
+    LUCKY_DEVIL,
+    JINX,
+    CLAMMY_CLOTHES,
+    CONSTANT_CLAMMY_CLOTHES,
+    ROBUST_STOMACH,
+    TOUGHNESS,
+    BABYSITTER,
+    HONEY_TRAP,
+    BANG_AND_BURN,
+    FLAPS_AND_SEALS,
+    TRADECRAFT,
+    OBSERVATION
+  };
+
+
+  //TODO: Wofuer ist das gut ? -> never used
+  enum class NPCEnum {
+    CAT,
+    JANITOR
+  };
+}
+
+
+#endif //LIBCOMMON_SRC_DATATYPES_CHARACTER_PROPERTYENUM_HPP_
